refactor: prep lint setup for upcoming diagnostics
